+++ labefood/src/routes/router.js
@@ -1,5 +1,5 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import { Header } from '../components';
+
 import {
     LoginPage,
     AddAdressPage,
@@ -15,8 +15,7 @@ import {
 export const Router = () => {
     return (
         <BrowserRouter>
-       <Header/>
-            <Routes>
+        <Routes>
                 <Route path='/login' element={<LoginPage />} />
                 <Route path='/signup' element={<SignUpPage />} />
                 <Route path='/feed' element={<FeedPage />} />
